+++ react/__tests__/DocumentField.test.tsx
@@ -1,16 +1,18 @@
 import * as React from 'react'
-import { render, fireEvent } from '@vtex/test-tools/react'
+import { render, fireEvent, wait } from '@vtex/test-tools/react'
 
 import DocumentField from '../DocumentField'
 
 describe('DocumentField', () => {
-  it('should correctly mask BRA documents', () => {
+  it('should correctly mask BRA documents', async () => {
     const Component = () => {
+      const [document, setDocument] = React.useState('11111111111')
       return (
         <DocumentField
           label="Document"
           documentType="cpf"
-          document="11111111111"
+          document={document}
+          onChange={(data: { document: string }) => setDocument(data.document)}
         />
       )
     }
@@ -19,10 +21,10 @@ describe('DocumentField', () => {
 
     const input = getByLabelText(/document/i)
 
-    expect(input).toHaveValue('111.111.111-11')
+    await wait(() => void expect(input).toHaveValue('111.111.111-11'))
   })
 
-  it('should clamp exceeding characters on change', () => {
+  it('should only mask value on field blur', () => {
     const Component: React.FC = () => {
       const [document, setDocument] = React.useState('')
       const [documentType, setDocumentType] = React.useState('cpf')
@@ -49,15 +51,111 @@ describe('DocumentField', () => {
 
     const input = getByLabelText(/document/i)
 
+    fireEvent.focus(input)
+
     fireEvent.change(input, { target: { value: '123456789' } })
 
-    expect(input).toHaveValue('123.456.789')
+    expect(input).toHaveValue('123456789')
 
     fireEvent.change(input, { target: { value: '12345678901' } })
 
-    expect(input).toHaveValue('123.456.789-01')
+    expect(input).toHaveValue('12345678901')
 
-    fireEvent.change(input, { target: { value: '123456789012' } })
+    fireEvent.blur(input)
+
+    expect(input).toHaveValue('123.456.789-01')
+  })
+
+  it("should not format if document isn't cpf", () => {
+    const AnotherDocComponent: React.FC = () => {
+      const [document, setDocument] = React.useState('')
+      const [documentType, setDocumentType] = React.useState('another-document')
+
+      const handleDocumentChange = (data: {
+        document: string
+        documentType: string
+      }) => {
+        setDocument(data.document)
+        setDocumentType(data.documentType)
+      }
+
+      return (
+        <DocumentField
+          label="Document"
+          document={document}
+          documentType={documentType}
+          onChange={handleDocumentChange}
+        />
+      )
+    }
+
+    const { getByLabelText } = render(<AnotherDocComponent />)
+
+    const input = getByLabelText(/document/i)
+
+    fireEvent.focus(input)
+
+    fireEvent.change(input, { target: { value: '123456789' } })
+
+    fireEvent.blur(input)
+
+    expect(input).toHaveValue('123456789')
+  })
+
+  it('should allow input of any special characters', () => {
+    const Component: React.FC = () => {
+      const [document, setDocument] = React.useState('')
+      const [documentType, setDocumentType] = React.useState('')
+
+      const handleDocumentChange = (data: {
+        document: string
+        documentType: string
+      }) => {
+        setDocument(data.document)
+        setDocumentType(data.documentType)
+      }
+
+      return (
+        <DocumentField
+          label="Document"
+          document={document}
+          documentType={documentType}
+          onChange={handleDocumentChange}
+        />
+      )
+    }
+
+    const { getByLabelText } = render(<Component />)
+
+    const input = getByLabelText(/document/i)
+
+    fireEvent.change(input, { target: { value: 'abc123.,[]&*()' } })
+
+    expect(input).toHaveValue('abc123.,[]&*()')
+  })
+
+  it('should be able to mask a partially masked BRA document', () => {
+    const ComponentWithPartialDoc = () => {
+      const [doc, setDoc] = React.useState('')
+      return (
+        <DocumentField
+          label="Document"
+          documentType="cpf"
+          document={doc}
+          onChange={(data: { document: string }) => setDoc(data.document)}
+        />
+      )
+    }
+
+    const { getByLabelText } = render(<ComponentWithPartialDoc />)
+
+    const input = getByLabelText(/document/i)
+
+    fireEvent.focus(input)
+
+    fireEvent.change(input, { target: { value: '123.45678901' } })
+
+    fireEvent.blur(input)
 
     expect(input).toHaveValue('123.456.789-01')
   })