+++ react/utils/validation.ts
@@ -1,11 +1,16 @@
+export const CPF_LENGTH = 11
 export const CPF_MASK = '999.999.999-99'
 
-export const unmaskCPF = (value: string) => value.replace(/\D/g, '')
+const unmaskCPF = (value: string) => value.replace(/\D/g, '')
 
 export const validateCPF = (value: string) => {
   const unmaskedValue = unmaskCPF(value)
 
-  if (unmaskedValue.length < 11) {
+  if (unmaskedValue.length !== CPF_LENGTH) {
+    return false
+  }
+
+  if (unmaskedValue.split('').every(digit => digit === unmaskedValue[0])) {
     return false
   }