+++ CHANGELOG.md
@@ -5,9 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Masking of the document will now occur only on the `blur` event.
+- The document input now allows any character.
 
 ## [0.1.0] - 2020-03-24
-
 ### Added
-
 - Initial implementation for `DocumentField` component.